LuscaUtils/requirements.txt
numpy
pillow
onnxruntime
opencv-python
onnx
